requirements.txt
pyspark
numpy
seaborn
matplotlib
from-root==1.0.2
requests
evidently
PyYAML==6.0.1
pymysql
apache-airflow
connexion[swagger-ui]
#tqdm
-e .
